+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.1
 - Fixed startup crash
+- Now Ancient Rarity only appears in the Affix Codex if Ancient Reforging is installed
 
 # 1.0.0 
 - Initial Release